requirements.txt
dash==2.9.3
dash-bootstrap-components==1.4.1
pandas==1.5.3
scikit-surprise==1.1.3
numpy==1.23.5
requests==2.28.2
python-dotenv==1.0.0
scikit-learn==1.2.2